refactor(version): 拆分VersionManagerModal为版本对比+备份恢复
