+++ frontend/src/components/instances/version/index.ts
@@ -1,0 +1,2 @@
+export { VersionComparisonTab } from './VersionComparisonTab';
+export { BackupRestoreTab } from './BackupRestoreTab';
